src/pyproject.toml
[tool.poetry]
name = "WABot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1.7"
djangorestframework = "^3.12.4"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
